requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
line-bot-sdk==3.4.0
python-multipart==0.0.6
python-dotenv==1.0.0